crud user

+++ src/main/java/com/example/tapchikhcn/repository/UserRepository.java
@@ -2,12 +2,14 @@ package com.example.tapchikhcn.repository;
 
 import com.example.tapchikhcn.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer > {
     UserEntity findByUsername(String username);
+
 
 }
